requirements.txt
anyio==4.4.0
certifi==2024.7.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-telegram-bot==21.4
sniffio==1.3.1
groq==0.9.0
asyncpg
psycopg2-binary
matplotlib
aiohttp